# api-spec.yaml
openapi: 3.1.0
info:
  title: Median Code API
  description: Production-ready FastAPI code generation and API entity management
    with Clerk JWT authentication.
  version: 1.0.0
paths:
...
  /v1/apis:
    get:
      tags:
      - APIs
      summary: List all APIs
      description: Retrieve all API definitions accessible to the authenticated user.
      operationId: list_apis_v1_apis_get
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiResponse'
                title: Response List Apis V1 Apis Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - APIs
      summary: Create a new API
      description: Create a new API definition.
      operationId: create_api_v1_apis_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/apis/{api_id}:
    get:
      tags:
      - APIs
      summary: Get API by ID
      description: Retrieve a specific API definition by its ID.
      operationId: get_api_v1_apis__api_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: api_id
        in: path
        required: true
        schema:
          type: string
          title: Api Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - APIs
      summary: Update API
      description: Update an existing API definition.
      operationId: update_api_v1_apis__api_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: api_id
        in: path
        required: true
        schema:
          type: string
          title: Api Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - APIs
      summary: Delete API
      description: Delete an API definition. This cascades to endpoints.
      operationId: delete_api_v1_apis__api_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: api_id
        in: path
        required: true
        schema:
          type: string
          title: Api Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/apis/{api_id}/generate:
    post:
      tags:
      - APIs
      summary: Generate FastAPI application
      description: Generates a complete FastAPI application for the specified API.
      operationId: generate_api_code_v1_apis__api_id__generate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: api_id
        in: path
        required: true
        schema:
          type: string
          title: Api Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateOptions'
              default:
                responsePlaceholders: true
                databaseEnabled: false
                cdkEnabled: false
                cdkCompute: lambda
      responses:
        '200':
          description: Successfully generated FastAPI application
          content:
            application/json:
              schema: {}
            application/zip:
              schema:
                type: string
                format: binary
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/endpoints:
    get:
      tags:
      - Endpoints
      summary: List all endpoints
      description: Retrieve all API endpoint definitions accessible to the authenticated
        user.
      operationId: list_endpoints_v1_endpoints_get
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiEndpointResponse'
                title: Response List Endpoints V1 Endpoints Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Endpoints
      summary: Create a new endpoint
      description: Create a new API endpoint definition.
      operationId: create_endpoint_v1_endpoints_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiEndpointCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEndpointResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/endpoints/{endpoint_id}:
    get:
      tags:
      - Endpoints
      summary: Get endpoint by ID
      description: Retrieve a specific API endpoint by its ID.
      operationId: get_endpoint_v1_endpoints__endpoint_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: endpoint_id
        in: path
        required: true
        schema:
          type: string
          title: Endpoint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEndpointResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Endpoints
      summary: Update endpoint
      description: Update an existing API endpoint definition.
      operationId: update_endpoint_v1_endpoints__endpoint_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: endpoint_id
        in: path
        required: true
        schema:
          type: string
          title: Endpoint Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiEndpointUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEndpointResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Endpoints
      summary: Delete endpoint
      description: Delete an API endpoint definition.
      operationId: delete_endpoint_v1_endpoints__endpoint_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: endpoint_id
        in: path
        required: true
        schema:
          type: string
          title: Endpoint Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ApiCreate:
      properties:
        namespaceId:
          type: string
          format: uuid
          title: Namespaceid
          examples:
          - 00000000-0000-0000-0000-000000000002
        title:
          type: string
          title: Title
          examples:
          - UserManagementApi
        version:
          type: string
          title: Version
          examples:
          - 1.0.0
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          default: ''
          examples:
          - API for managing user accounts
        baseUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Baseurl
          default: ''
          examples:
          - /api/v1
        serverUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Serverurl
          default: ''
          examples:
          - https://api.example.com
      type: object
      required:
      - namespaceId
      - title
      - version
      title: ApiCreate
      description: 'Request schema for creating an API.


        :ivar namespace_id: Namespace this API belongs to.

        :ivar title: API title in PascalCase (used as project identifier).

        :ivar version: Semantic version string.

        :ivar description: API description.

        :ivar base_url: Base path for all endpoints.

        :ivar server_url: Full server URL.'
    ApiEndpointCreate:
      properties:
        apiId:
          type: string
          format: uuid
          title: Apiid
          examples:
          - 00000000-0000-0000-0005-000000000001
        method:
          type: string
          enum:
          - GET
          - POST
          - PUT
          - PATCH
          - DELETE
          title: Method
          examples:
          - GET
        path:
          type: string
          title: Path
          examples:
          - /users/{user_id}
        description:
          type: string
          title: Description
          examples:
          - Get user by ID
        tagName:
          anyOf:
          - type: string
          - type: 'null'
          title: Tagname
          examples:
          - Users
        pathParams:
          items:
            $ref: '#/components/schemas/PathParamSchema'
          type: array
          title: Pathparams
        queryParamsObjectId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Queryparamsobjectid
          examples:
          - 00000000-0000-0000-0007-000000000001
        objectId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Objectid
          examples:
          - 00000000-0000-0000-0007-000000000002
        useEnvelope:
          type: boolean
          title: Useenvelope
          examples:
          - true
        responseShape:
          type: string
          enum:
          - object
          - list
          title: Responseshape
          examples:
          - object
      type: object
      required:
      - apiId
      - method
      - path
      - description
      - pathParams
      - useEnvelope
      - responseShape
      title: ApiEndpointCreate
      description: 'Request schema for creating an endpoint.


        :ivar api_id: API this endpoint belongs to.

        :ivar method: HTTP method.

        :ivar path: URL path with optional parameters.

        :ivar description: Endpoint description.

        :ivar tag_name: Optional tag name (must exist in the parent API''s tags).

        :ivar path_params: Path parameters referencing field definitions.

        :ivar query_params_object_id: Optional query params object reference.

        :ivar object_id: Optional object reference for request/response body.

        :ivar use_envelope: Whether to wrap response in envelope.

        :ivar response_shape: Response shape (object or list).'
    ApiEndpointResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          examples:
          - 00000000-0000-0000-0004-000000000001
        apiId:
          type: string
          format: uuid
          title: Apiid
          examples:
          - 00000000-0000-0000-0005-000000000001
        method:
          type: string
          enum:
          - GET
          - POST
          - PUT
          - PATCH
          - DELETE
          title: Method
          examples:
          - GET
        path:
          type: string
          title: Path
          examples:
          - /users/{user_id}
        description:
          type: string
          title: Description
          examples:
          - Retrieve user by ID
        tagName:
          anyOf:
          - type: string
          - type: 'null'
          title: Tagname
          examples:
          - Users
        pathParams:
          items:
            $ref: '#/components/schemas/PathParamSchema'
          type: array
          title: Pathparams
        queryParamsObjectId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Queryparamsobjectid
          examples:
          - 00000000-0000-0000-0007-000000000001
        objectId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Objectid
          examples:
          - 00000000-0000-0000-0007-000000000002
        useEnvelope:
          type: boolean
          title: Useenvelope
          examples:
          - true
        responseShape:
          type: string
          enum:
          - object
          - list
          title: Responseshape
          examples:
          - object
      type: object
      required:
      - id
      - apiId
      - method
      - path
      - description
      - useEnvelope
      - responseShape
      title: ApiEndpointResponse
      description: 'Response schema for endpoint data.


        :ivar id: Unique identifier for the endpoint.

        :ivar api_id: API this endpoint belongs to.

        :ivar method: HTTP method.

        :ivar path: URL path.

        :ivar description: Endpoint description.

        :ivar tag_name: Tag name (references a tag in the parent API).

        :ivar path_params: Path parameters referencing field definitions.

        :ivar query_params_object_id: Query params object reference.

        :ivar object_id: Object reference for request/response body.

        :ivar use_envelope: Whether response is wrapped in envelope.

        :ivar response_shape: Response shape.'
    ApiEndpointUpdate:
      properties:
        apiId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Apiid
          examples:
          - 00000000-0000-0000-0005-000000000001
        method:
          anyOf:
          - type: string
            enum:
            - GET
            - POST
            - PUT
            - PATCH
            - DELETE
          - type: 'null'
          title: Method
          examples:
          - POST
        path:
          anyOf:
          - type: string
          - type: 'null'
          title: Path
          examples:
          - /users
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - Updated endpoint description
        tagName:
          anyOf:
          - type: string
          - type: 'null'
          title: Tagname
        pathParams:
          anyOf:
          - items:
              $ref: '#/components/schemas/PathParamSchema'
            type: array
          - type: 'null'
          title: Pathparams
        queryParamsObjectId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Queryparamsobjectid
        objectId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Objectid
        useEnvelope:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Useenvelope
        responseShape:
          anyOf:
          - type: string
            enum:
            - object
            - list
          - type: 'null'
          title: Responseshape
      type: object
      title: ApiEndpointUpdate
      description: 'Request schema for updating an endpoint.


        :ivar api_id: Updated API reference.

        :ivar method: Updated HTTP method.

        :ivar path: Updated URL path.

        :ivar description: Updated description.

        :ivar tag_name: Updated tag name.

        :ivar path_params: Updated path parameters.

        :ivar query_params_object_id: Updated query params object reference.

        :ivar object_id: Updated object reference for request/response body.

        :ivar use_envelope: Updated envelope setting.

        :ivar response_shape: Updated response shape.'
    ApiResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          examples:
          - 00000000-0000-0000-0005-000000000001
        namespaceId:
          type: string
          format: uuid
          title: Namespaceid
          examples:
          - 00000000-0000-0000-0000-000000000001
        title:
          type: string
          title: Title
          examples:
          - User Management API
        version:
          type: string
          title: Version
          examples:
          - 1.0.0
        description:
          type: string
          title: Description
          examples:
          - Endpoints for user management
        baseUrl:
          type: string
          title: Baseurl
          examples:
          - /api/v1
        serverUrl:
          type: string
          title: Serverurl
          examples:
          - https://api.example.com
        createdAt:
          type: string
          format: date-time
          title: Createdat
          examples:
          - '2026-01-25T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
          examples:
          - '2026-01-25T10:30:00Z'
      type: object
      required:
      - id
      - namespaceId
      - title
      - version
      - description
      - baseUrl
      - serverUrl
      - createdAt
      - updatedAt
      title: ApiResponse
      description: 'Response schema for API data.


        :ivar id: Unique identifier for the API.

        :ivar namespace_id: Namespace this API belongs to.

        :ivar title: API title.

        :ivar version: Semantic version string.

        :ivar description: API description.

        :ivar base_url: Base path for all endpoints.

        :ivar server_url: Full server URL.

        :ivar created_at: Creation timestamp.

        :ivar updated_at: Last update timestamp.'
    ApiUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
          examples:
          - UpdatedApiTitle
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
          examples:
          - 2.0.0
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - Updated API description
        baseUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Baseurl
          examples:
          - /api/v2
        serverUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Serverurl
          examples:
          - https://api.example.com
      type: object
      title: ApiUpdate
      description: 'Request schema for updating an API.


        :ivar title: Updated API title in PascalCase.

        :ivar version: Updated version string.

        :ivar description: Updated API description.

        :ivar base_url: Updated base path.

        :ivar server_url: Updated server URL.'
...
    GenerateOptions:
      properties:
        responsePlaceholders:
          type: boolean
          title: Responseplaceholders
          default: true
        databaseEnabled:
          type: boolean
          title: Databaseenabled
          default: false
        cdkEnabled:
          type: boolean
          title: Cdkenabled
          default: false
        cdkCompute:
          type: string
          enum:
          - lambda
          - ecs
          title: Cdkcompute
          default: lambda
      type: object
      title: GenerateOptions
      description: "Options for code generation passed to POST /v1/apis/{api_id}/generate.\n\
        \n:ivar response_placeholders: Generate placeholder response bodies.\n:ivar\
        \ database_enabled: Generate database support (SQLAlchemy, Alembic, Docker\
        \ Compose).\n:ivar cdk_enabled: Generate CDK infrastructure files.\n:ivar\
        \ cdk_compute: Compute platform for CDK \u2014 'lambda' or 'ecs'."
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer